rename reflectionProbeManager
